// ClientApp/src/app/record-room/models/DegitalArchieve.ts
export interface DegitalArchieve {
    degitalArchieveId: number;
    departmentNameId: number;
    airCraftNameId: number;
    degitalArchieveDocTypeId:number;
    name: number;
    dateOfLastRev:number;
    doc:number;
    remarks: string;
    departmentName: string;
    menuPosition:Date;
    isActive: boolean;
   
}